requirements.txt
numpy==1.22.1
python-docx==0.8.11
mistletoe==0.8.1
regex==2022.1.18
pandas==1.3.5
openpyxl==3.0.9
auto-py-to-exe==2.15.0